+++ routes/genres.js
@@ -2,6 +2,7 @@ const {validate, Genre} = require('../models/genre');
 const express = require('express');
 const router = express.Router();
 const mongoose = require('mongoose');
+const validateId = require('../utils/validateId');
 
 mongoose.connect('mongodb://localhost/vidly')
     .then(() => console.log('Connected to MongoDB'))
@@ -12,6 +13,9 @@ router.get('/', async (req, res) => {
 });
 
 router.get('/:id', async (req, res) => {
+    const errorId = validateId(req.params.id);
+    if(errorId) return res.status(400).send('Invalid Genre ID');
+
     const genre = await Genre.findById(req.params.id);
     if(!genre) return res.status(404).send(`Genre with id ${req.params.id} not found`);
     res.send(genre);
@@ -19,7 +23,6 @@ router.get('/:id', async (req, res) => {
 
 router.post('/', async (req, res) => {
     const { error } = validate(req.body);
-
     if (error) return res.status(400).send(error.details[0].message);
 
     let genre = new Genre({name: req.body.name});
@@ -28,6 +31,9 @@ router.post('/', async (req, res) => {
 });
 
 router.put('/:id', async (req, res) => {
+    let errorId = validateId(req.params.id);
+    if(errorId) return res.status(400).send('Invalid Genre ID');
+
     const { error } = validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
@@ -38,6 +44,9 @@ router.put('/:id', async (req, res) => {
 });
 
 router.delete('/:id', async (req, res) => {
+    const errorId = validateId(req.params.id);
+    if(errorId) return res.status(400).send('Invalid Genre ID');
+
     const genre = await Genre.findByIdAndDelete(req.params.id);
     if(!genre) return res.status(404).send(`Genre with id ${req.params.id} is not found!!!`);
     res.send(genre);
